Migrate code for merge
